+++ src/App.tsx
@@ -1,20 +1,20 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 
-import { Home } from './components/pages/Home/';
-import { AddBook } from './components/pages/AddBook';
-import { EditBook } from './components/pages/EditBook';
-import { Login } from './components/pages/Login';
-import { SignUp } from './components/pages/SignUp';
+import { HomeRouter } from './components/pages/Home/homeRouter';
+import { AddBookRouter } from './components/pages/AddBook/addBookRouter';
+import { EditBookRouter } from './components/pages/EditBook/editBookRouter';
+import { LoginRouter } from './components/pages/Login/loginRouter';
+import { SignUpRouter } from './components/pages/SignUp/signUpRouter';
 
 function App() {
   return (
     <Switch>
-      <Route exact path="/" component={Home} />
-      <Route exact path="/add" component={AddBook} />
-      <Route exact path="/edit" component={EditBook} />
-      <Route exact path="/login" component={Login} />
-      <Route exact path="/sign-up" component={SignUp} />
+      <Route exact path="/" component={HomeRouter} />
+      <Route exact path="/add" component={AddBookRouter} />
+      <Route exact path="/edit" component={EditBookRouter} />
+      <Route exact path="/login" component={LoginRouter} />
+      <Route exact path="/sign-up" component={SignUpRouter} />
     </Switch>
   );
 }
