+++ src/components/pages/EditBook/editBookRouter.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { HeaderComponent } from '../../molecules/Header/index';
 import { BookScreenComponent } from '../../organisms/BookScreen/index';
 import Paper from '@material-ui/core/Paper';
-import { PaperContainer, useStyles } from './style';
+import { PaperContainer, useStyles } from '../AddBook/style';
 
-export function AddBook() {
+export function EditBookRouter() {
   const classes = useStyles();
 
   return (
@@ -12,7 +12,7 @@ export function AddBook() {
       <HeaderComponent headerType={'main'}></HeaderComponent>
       <PaperContainer>
         <Paper elevation={8} className={classes.paper}>
-          <BookScreenComponent bookMode="add"></BookScreenComponent>
+          <BookScreenComponent bookMode="edit"></BookScreenComponent>
         </Paper>
       </PaperContainer>
     </>
