+++ src/components/pages/Login/loginRouter.tsx
@@ -3,7 +3,7 @@ import { AuthComponent } from '../../organisms/Auth/index';
 import Paper from '@material-ui/core/Paper';
 import { PaperContainer, Title } from './style';
 
-export function Login() {
+export function LoginRouter() {
   return (
     <>
       <Title>Book Manager</Title>
